[PoC] Add messages for 3 languages and demo with Button
This commit adds a message files containing objects per lang code
with standard sets of messages that can be used in stories. This
is demonstrated in the configurable Button story. When the user
changes the language in the toolbar, the text sample switches to
that language. Some considerations:

- The user needs to be able to both switch between languages in the
toolbar and see the text change in the story, and set their own
custom text in the story controls
- The way I've done it here is simple from a configuration standpoint
but takes a bit of extra work on the individual story side
- This system is built on a small set of text samples that must be
completely generic so they can be reused across components. However,
this might not be as clear to users looking for component demos than
component-specific sample text
- It might be better from a DX standpoint to be able to use an i18n
plugin similar to what we've got in MediaWiki (or the one Santhosh
built) that's based on banana-i18n. We probably don't need the full
functionality of banana-i18n (like plurals and gender), but it
would be nice to hook up a plugin so we can avoid the extra lines
of code per-story.